+++ smos-report-cursor/src/Smos/Cursor/Report/Stuck.hs
@@ -7,30 +7,26 @@ import Conduit
 import Cursor.Simple.List.NonEmpty
 import qualified Data.Conduit.Combinators as C
 import qualified Data.List.NonEmpty as NE
-import qualified Data.Map as M
-import Data.Ord
 import Data.Time
 import Data.Validity
 import Data.Validity.Path ()
 import GHC.Generics
 import Lens.Micro
 import Path
-import Smos.Cursor.Report.Entry
-import Smos.Cursor.SmosFile
 import Smos.Data
-import Smos.Report.Archive
 import Smos.Report.Config
-import Smos.Report.Filter
-import Smos.Report.Period
 import Smos.Report.ShouldPrint
 import Smos.Report.Streaming
 import Smos.Report.Stuck
 
-produceStuckReportCursor :: TimeZone -> HideArchive -> ShouldPrint -> DirectoryConfig -> IO StuckReportCursor
-produceStuckReportCursor tz ha sp dc = produceReport ha sp dc $ stuckReportCursorConduit tz
+produceStuckReportCursor :: TimeZone -> ShouldPrint -> DirectoryConfig -> IO StuckReportCursor
+produceStuckReportCursor tz sp dc = runConduit $ streamSmosProjects sp dc .| stuckReportCursorConduit tz
 
 stuckReportCursorConduit :: Monad m => TimeZone -> ConduitT (Path Rel File, SmosFile) void m StuckReportCursor
-stuckReportCursorConduit tz = StuckReportCursor . fmap makeNonEmptyCursor . NE.nonEmpty . sortStuckEntries <$> (C.concatMap (uncurry $ makeStuckReportEntry tz) .| sinkList)
+stuckReportCursorConduit tz =
+  StuckReportCursor . fmap makeNonEmptyCursor . NE.nonEmpty
+    . sortStuckEntries
+    <$> (C.concatMap (uncurry $ makeStuckReportEntry tz) .| sinkList)
 
 newtype StuckReportCursor = StuckReportCursor
   { stuckReportCursorNonEmptyCursor :: Maybe (NonEmptyCursor StuckReportEntry)